Add e2e tests using chainsaw (#590)
* Add e2e tests using chainsaw

We attempted to roll our own E2E framework in https://github.com/kro-run/kro/pull/290
The []steps{input,outputs} model the DIY framework match is a direct
match to the []steps{apply, assert} model of chainsaw.

So we can port the tests from the other PR to this as is.
The go-test code in the DIY framework would be translate to
`chainsaw-test.yaml` for each test scenario (folder)

Ported over these tests:
* TestAutoscaledDeploymentRGD -> multiple-resource
* SimpleDeployment -> simple-deployment

* Addressing review comments